refactor macro-level grammar trees into own file
